[rename] 모임 상세 페이지 관련 컴포넌트 파일 이동 및 이름 수정, apiPaths readonly 설정
