+++ src/components/gatherings/gatheringDetail/Footer.tsx
