+++ src/components/gatherings/gatheringDetail/Information.tsx
@@ -1,4 +1,4 @@
-import { Participant } from '@/app/gatherings/detail/[id]/ClientPage';
+import { Participant } from '@/components/gatherings/detail/ClientPage';
 import { formatDate, formatTime } from '@/components/shared/utils/format';
 import { Gathering } from '@/types/gatherings';
 import { Check, UserRoundCheck } from "lucide-react"
