+++ src/components/gatherings/gatheringDetail/Thumbnail.tsx
