+++ src/components/gatherings/gatheringDetail/loading/InformationLoading.tsx
