+++ src/components/gatherings/gatheringDetail/loading/ThumbnailLoading.tsx
